requirements.txt
databricks-sdk>=0.40.0
databricks-sql-connector>=3.0.0
pandas>=1.5.0
plotly>=5.0.0
streamlit>=1.28.0
python-dotenv>=0.19.0
scipy>=1.9.0
numpy>=1.21.0